feat(AddGoalToModel): Multi Fix
- Added RegistryType.cs (enum) to replace the IndicatorType.cs.
- Replaced Type by RegistriesType in Indicator.cs.
- Replaced type by registriesType in indicator.ts.
- Replaced Type by RegistriesType in DataContext.cs (to do the correct mapping).
- Modified registry.form.component.ts to make use of registryType.ts instead of indicatorsType.cs.
- Modified registry-form.component.html to make use of the indicator.registriesType instead of indicator.registries[0].discriminator and deleted a comment and an incorrect field used for the default registries.
- Changed on registry-editor.component.html the necessary values for every type of input according to the new enum (RegistryType) and the name of the variable type to registriesType.
- Changed on registry.editor.component.ts the name of the parameter type to registriesType.
- Modified the name of the variable type to registriesType , deleted the import of IndicatorType and deleted an unnecessary comment on indicator-detail.component.cs.
- Modified indicator.type by indicator.registiresType and the conditions in the *ngIf (according the new enum RegistryType) on indicator-detail.component.html.
- Fixed indentation problems (tabs instead of 2 spaces) in indicator.service.ts and deleted the import of indicatorType.ts (deleted).
- Changed Quantity in QuantityRegistry.cs to long.
- Changed ActivityRegistry according the required data.
- Changed QuantityIndicatorCalculator.cs, now ActivityRegistry is calculated by its quantity parameter and to throws an error in case that it's passed an incorrect registry to it and the sum is a long (because quantity is long).
- Added an IndicatorID in Goal.cs (to be able to access to the indicator parent of the goal).
- Changed Percent to double in PercentRegistry.cs.
- Changed PercentIndicatorCalculator.cs to only allow a PercentRegistry (by now) to do the respective calculation if the registry is of a different type, it throws an exception.
- Modified the Get by id methods in IndicatorController.cs and IndicatorGroupsControlles.cs to improve the performance (before it loaded the whole data of all the Indicators/IndicatorGroups, now only loads the data of the specific Indicator/IndicatorGroup).
- Modified the constructor of registry.ts, now quantity starts with 0 instead of -1.
- Deleted indicatorType.cs (replaced by RegistryType).
- Modified the DbInitializer.cs to specify the registriesType (for the indicators that don't use the default) and added mock data for the goals (only in the indicators of the group 1).

+++ Models/PercentRegistry.cs
@@ -8,7 +8,7 @@ namespace think_agro_metrics.Models
 {
     public class PercentRegistry : Registry
     {
-        public int Percent { get; set; }
+        public double Percent { get; set; }
 
         public PercentRegistry() { }
     }
